Refactor: Remove callback
